plat-199 refactored code and fixed test cases
