+++ src/main/java/gov/va/ascent/framework/security/VAServiceWss4jSecurityInterceptor.java
@@ -9,8 +9,6 @@ import org.springframework.ws.soap.security.wss4j2.Wss4jSecuritySecurementExcept
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 
-import gov.va.ascent.framework.security.BEPWebServiceUtil;
-
 /**
  * Extension of the Wss4jSecurityInterceptor to inject the VA application headers into the WSS security element
  * 
@@ -20,37 +18,37 @@ import gov.va.ascent.framework.security.BEPWebServiceUtil;
 public class VAServiceWss4jSecurityInterceptor extends Wss4jSecurityInterceptor {
 
 	/** The namespace constant. */
-	protected static final String VA_NS = "http://vbawebservices.vba.va.gov/vawss";
+	public static final String VA_NS = "http://vbawebservices.vba.va.gov/vawss";
 
 	/** The prefix constant */
-	protected static final String VA_PREFIX = "vaws:";
+	public static final String VA_PREFIX = "vaws:";
 
 	/** The headers node constant */
-	protected static final String VA_SERVICE_HEADERS = "VaServiceHeaders";
+	public static final String VA_SERVICE_HEADERS = "VaServiceHeaders";
 
 	/** The application name node constant */
-	protected static final String VA_APPLICATION_NAME = "applicationName";
+	public static final String VA_APPLICATION_NAME = "applicationName";
 
 	/** The Constant CLIENT_MACHINE. */
-	protected static final String CLIENT_MACHINE = "CLIENT_MACHINE";
+	public static final String CLIENT_MACHINE = "CLIENT_MACHINE";
 
 	/** The Constant STN_ID. */
-	protected static final String STN_ID = "STN_ID";
+	public static final String STN_ID = "STN_ID";
 
 	/** The Constant EXTERNAL_UID. */
-	protected static final String EXTERNAL_UID = "ExternalUid";
+	public static final String EXTERNAL_UID = "ExternalUid";
 
 	/** The Constant EXTERNAL_KEY. */
-	protected static final String EXTERNAL_KEY = "ExternalKey";
+	public static final String EXTERNAL_KEY = "ExternalKey";
 
 	/** The Constant for when a userid cannot be obtained. */
-	protected static final String EXTERNAL_UID_DEFAULT = "EVSS";
+	public static final String EXTERNAL_UID_DEFAULT = "EVSS";
 
 	/** The va client machine. */
 	private String clientMachine;
 
 	/** The va station id. */
-	private String stationId = "281";	//RR: Temporary, to be REMOVED
+	private String stationId;
 
 	/** The va application name. */
 	private String vaApplicationName;
@@ -62,7 +60,7 @@ public class VAServiceWss4jSecurityInterceptor extends Wss4jSecurityInterceptor 
 	 * org.springframework.ws.soap.SoapMessage, org.springframework.ws.context.MessageContext)
 	 */
 	@Override
-	protected final void secureMessage(final SoapMessage soapMessage, final MessageContext messageContext) {
+	public final void secureMessage(final SoapMessage soapMessage, final MessageContext messageContext) {
 		super.secureMessage(soapMessage, messageContext);
 
 		// Build the VA Header Element
